Refactor all postgres tables; start work on sqlite
